requirements.txt
# Core
python-dotenv>=1.0.0

# Web framework
fastapi>=0.95.0
uvicorn>=0.21.0

# Utilities
pydantic>=1.10.0
loguru>=0.6.0

# Text processing
nltk>=3.8.1
markovify>=0.9.4

# MIDI processing
mido>=1.2.10
python-rtmidi>=1.4.9
music21>=9.1.0
midiutil>=1.2.1

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0
numpy>=1.24.0
scipy>=1.10.0

# Audio format conversion
pydub>=0.25.1

# GGUF model support
llama-cpp-python>=0.2.0

# Transformers (for rinna model fallback and fine-tuning)
transformers>=4.30.0
torch>=2.0.0
datasets>=2.14.0

# ML analysis
scikit-learn>=1.3.0
